Add watchdog timers + use watchdog timer in modem code (#567)
* thread: Add one-shot timers

A one-shot timer is a timer that will trigger an action (through a
pre-registered callback) after a timeout expires. The code using a
one-shot timer will typically reset it periodically to prevent the
action from happening, kind of like a watchdog.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>

* modem: Poll modem properly

Instead of having a convoluted method to handle timeouts and then
busy-wait on a timeout counter, poll the modem the proper way and sleep
until the proper bits are set or cleared.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>

* modem: Make some global symbols private

Some global symbols aren't used anywhere else. Therefore they can be
made private.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>

* modem: Use one-shot timer to handle timeouts

Instead of directly using the timer API and reserving the TIMER UNIT N°1
just for the modem, use a one-shot timer to handle timeouts.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>

---------

Signed-off-by: Paul Cercueil <paul@crapouillou.net>