fs_dcload: Set errno on error in dcload_stat() (#642)
The fs_dcload code doesn't seem to set errno anywhere. It means that on
error, the calling code will get whatever errno value was set by
anything that came before it.

In dcload_stat(), if the PC communication could not be done, set errno
to ENOENT - "No such file or directory". This will cause the upper
layers to not attempt to open the directory.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>