Create romdisks with bin2c instead of bin2o (#641)
* bin2c: Fix invalid generated code

The generated C was not valid as it was missing a semicolon on one of
its lines.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>

* romdisks: Constify __kos_romdisk

The romdisk data isn't supposed to be modified, so modify the
__kos_romdisk pointer to add the "const" specifier.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>

* romdisks: Create romdisks using bin2c instead of bin2o

bin2o creates a romdisk.o from a romdisk.img using the linker. On the
other hand, bin2c creates a C file with a char array that contains the
whole romdisk.img.

The problem with the first approach is that the romdisk.o generated has
technically not been compiled. This is a problem when building a project
with LTO, as the linker will refuse to perform whole-program
optimizations if some of the object files haven't been compiled with
LTO.

It then issues the following warning:
  ld: warning: incremental linking of LTO and non-LTO objects; using
  -flinker-output=nolto-rel which will bypass whole program optimization

Creating romdisks with bin2c then allows the linker to perform
whole-program optimizations.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>