Make FD_SETSIZE configurable (#703)
* Reduce FD_SETSIZE from 1024 => 64

The maximum amount of files that can be opened at once is 64.  We can save 11.25 KB if we reduce FD_SETSIZE to 64.

* Allow FD_SETSIZE value to be overwritten

Move all the #define FD_SETSIZE definitions to one file, opts.h.  Add an option to environ.sh to allow FD_SETSIZE to be configured. Add an printf message if opening a file fails and FD_SETSIZE is not set to the default.

* Updated documenation comment