thread: Add priority boosting system
Introduce a new 'real_prio' field into the thread structure, which
contains the static priority set during the thread creation or with
thd_set_prio().

The 'prio' field now represents the dynamic priority. When a
low-priority thread A holds a lock, and a high-priority thread B
requests it, the dynamic priority of A will be temporarily raised to the
dynamic priority of B, until A releases the lock, in which case it falls
back to its static priority.

The point of this setup is to avoid starvation, where a low-priority
thread holds a resource that is needed by a high-priority thread, and
won't release it because it rarely gets preempted.

This is not yet perfect, as we can imagine a setup where a low-priority
thread A holds two locks, each one requested by different high-priority
tasts B and C; A would get boosted to B's dynamic priority, and when
releasing the lock, reset to its original low priority, causing thread C
to starve.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>