Speed up building KOS and examples (#573)
* examples: Allow building all examples in parallel

Instead of building all the examples sequentially in a single thread,
build all of them in parallel.

The Makefile will first generate a list of all the sub-folders that
contain a Makefile, and all of them will be added to the list of phony
rules.

The default target will then depend on that sub-directory list, running
Make recursively on each of those subdirectories, also using the
special $(MAKE) variable so that the same jobserver is used.

Because of that, all the Makefiles used only for recursion are useless
now and can be dropped.

On my PC (i7-10850H 12-core), when building with LTO running
'time make -j12 -C examples/dreamcast/' used to take about 8min 15s, and
now takes about 1min 24s, while compiling more examples, as previously
the gldc/ examples were not compiled.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>

* utils: Use top-level Makefile rules for subdirs

Instead of rolling our own rules for handling the sub-directories,
include the top-level Makefile.rules and add a dependency on the
"subdirs" rule.

This actually speeds up the build quite a bit, as the default rule uses
$(MAKE) instead of $(KOS_MAKE), which means that the "make" program will
be able to use its jobserver.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>

* addons: Allow building all addons in parallel

Just like for the examples, generate the list of sub-directories using a
Makefile rule; and build the sub-directories using the general "subdirs"
rule, which will be much faster as the "make" program will be able to
use its jobserver.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>

* Makefile: Allow building in parallel

Build the sub-directories using the general "subdirs" rule, which will
be much faster as the "make" program will be able to  use its jobserver.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>

* examples: Allow examples to override the compatible KOS subarchs

The examples can override KOS_BUILD_SUBARCHS in their Makefile, to
override the list of compatible subarchs.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>

* examples: Override build subarchs for examples not Naomi-compatible

Some examples are not compatible with the Naomi, because they try to use
hardware that the Naomi does not have (e.g. the modem).

For those examples, set KOS_BUILD_SUBARCHS to "pristine", so that they
will only build for the "pristine" (aka. "dreamcast") sub-architecture.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>

---------

Signed-off-by: Paul Cercueil <paul@crapouillou.net>